+++ website/docs/data-display/area-chart/examples/area-custom-line.jsx
@@ -19,7 +19,7 @@ export default () => {
     .domain([0, 10]);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label='Example chart with a custom line'>
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
       </YAxis>