+++ website/docs/data-display/bar-horizontal/examples/bar-label.jsx
@@ -19,7 +19,7 @@ export default () => {
     .paddingOuter(0.2);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label="average cost of goods by category">
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
       </YAxis>