+++ website/docs/data-display/line-chart/examples/legend.jsx
@@ -87,7 +87,7 @@ export default () => {
             );
           })}
         </Flex>
-        <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label="chart comparing data by category">
+        <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
           <YAxis>
             <YAxis.Ticks ticks={yScale.ticks(4)} />
             <YAxis.Grid ticks={yScale.ticks(4)} />