+++ website/docs/data-display/line-chart/examples/line-curve.jsx
@@ -19,7 +19,7 @@ export default () => {
     .domain([0, 10]);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label="basic example of using line chart">
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
         <YAxis.Grid />