+++ website/docs/data-display/radar-chart/examples/fill.jsx
@@ -9,7 +9,7 @@ export default () => {
   const scale = scaleLinear().domain([0, 10]);
 
   return (
-    <Plot data={data} width={width} height={height} label="basic example of radar chart with transparent background">
+    <Plot data={data} width={width} height={height}>
       <Radar scale={scale}>
         <Radar.Axis dataKey='categories'>
           <Radar.Axis.Ticks />