+++ website/docs/data-display/radial-tree-chart/examples/radial-tree-multiline-text.jsx
@@ -40,7 +40,6 @@ export default () => {
       width={width}
       height={height}
       textSize={textSize}
-      label="example of radial chart with multiline text"
     >
       <RadialTree color="#AB6CFE">
         <RadialTree.Radian>