+++ website/docs/data-display/stacked-bar-chart/examples/bar-stack.jsx
@@ -20,7 +20,7 @@ export default () => {
     .domain([0, 20]);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label="chart of average cost of goods comparison by category">
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
         <YAxis.Grid />