+++ semcore/chart/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [5.3.21] - 2023-04-14
+## [5.3.20] - 2023-04-03
 
 ### Changed
 
-- Version patch update due to children dependencies update (`@semcore/utils` [3.50.0 ~> 3.50.2]).
+- Version patch update due to children dependencies update (`@semcore/checkbox` [6.4.9 ~> 6.4.10]).
 
 ## [5.3.9] - 2023-02-09