+++ semcore/dropdown/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [3.4.23] - 2023-04-14
+## [3.4.22] - 2023-04-03
 
 ### Changed
 
-- Version patch update due to children dependencies update (`@semcore/popper` [4.17.1 ~> 4.17.2], `@semcore/utils` [3.50.0 ~> 3.50.2]).
+- Version patch update due to children dependencies update (`@semcore/popper` [4.16.12 ~> 4.17.0]).
 
 ## [3.4.21] - 2023-03-28